Merge 4.2.0 master

+++ modules/tile-converter/src/3d-tiles-converter/3d-tiles-converter.ts
@@ -4,7 +4,7 @@ import type {
   NodeReference,
   I3STilesetHeader
 } from '@loaders.gl/i3s';
-import type {Tiles3DTileJSON} from '@loaders.gl/3d-tiles';
+import type {Tile3DBoundingVolume, Tiles3DTileJSON} from '@loaders.gl/3d-tiles';
 
 import {join} from 'path';
 import process from 'process';
@@ -24,8 +24,11 @@ import {WorkerFarm} from '@loaders.gl/worker-utils';
 import {BROWSER_ERROR_MESSAGE} from '../constants';
 import B3dmConverter, {I3SAttributesData} from './helpers/b3dm-converter';
 import {I3STileHeader} from '@loaders.gl/i3s/src/types';
-import {loadI3SContent} from './helpers/load-i3s';
+import {getNodeCount, loadFromArchive, loadI3SContent, openSLPK} from './helpers/load-i3s';
 import {I3SLoaderOptions} from '@loaders.gl/i3s/src/i3s-loader';
+import {ZipFileSystem} from '../../../zip/src';
+import {ConversionDump, ConversionDumpOptions} from '../lib/utils/conversion-dump';
+import {Progress} from '../i3s-converter/helpers/progress';
 
 const I3S = 'I3S';
 
@@ -41,6 +44,7 @@ export default class Tiles3DConverter {
   sourceTileset: I3STilesetHeader | null;
   attributeStorageInfo?: AttributeStorageInfo[] | null;
   workerSource: {[key: string]: string} = {};
+  slpkFilesystem: ZipFileSystem | null = null;
   loaderOptions: I3SLoaderOptions = {
     _nodeWorkers: true,
     reuseWorkers: true,
@@ -52,6 +56,8 @@ export default class Tiles3DConverter {
       workerUrl: './modules/i3s/dist/i3s-content-worker-node.js'
     }
   };
+  conversionDump: ConversionDump;
+  progress: Progress;
 
   constructor() {
     this.options = {};
@@ -62,6 +68,8 @@ export default class Tiles3DConverter {
     this.sourceTileset = null;
     this.attributeStorageInfo = null;
     this.workerSource = {};
+    this.conversionDump = new ConversionDump();
+    this.progress = new Progress();
   }
 
   /**
@@ -73,29 +81,53 @@ export default class Tiles3DConverter {
    * @param options.egmFilePath location of *.pgm file to convert heights from ellipsoidal to gravity-related format
    * @param options.maxDepth The max tree depth of conversion
    */
+  // eslint-disable-next-line complexity, max-statements
   public async convert(options: {
     inputUrl: string;
     outputPath: string;
     tilesetName: string;
     maxDepth?: number;
     egmFilePath: string;
-  }): Promise<any> {
+    inquirer?: Promise<unknown>;
+    analyze?: boolean;
+  }): Promise<string | undefined> {
     if (isBrowser) {
-      console.log(BROWSER_ERROR_MESSAGE);
+      console.log(BROWSER_ERROR_MESSAGE); // eslint-disable-line no-console
       return BROWSER_ERROR_MESSAGE;
     }
-    const {inputUrl, outputPath, tilesetName, maxDepth, egmFilePath} = options;
+    const {inputUrl, outputPath, tilesetName, maxDepth, egmFilePath, inquirer, analyze} = options;
     this.conversionStartTime = process.hrtime();
-    this.options = {maxDepth};
+    this.options = {maxDepth, inquirer};
 
     console.log('Loading egm file...'); // eslint-disable-line
     this.geoidHeightModel = await load(egmFilePath, PGMLoader);
     console.log('Loading egm file completed!'); // eslint-disable-line
 
-    this.sourceTileset = await load(inputUrl, I3SLoader, this.loaderOptions);
+    this.slpkFilesystem = await openSLPK(inputUrl);
+
+    let preprocessResult = true;
+    if (analyze || this.slpkFilesystem) {
+      preprocessResult = await this.preprocessConversion();
+      if (!preprocessResult || analyze) {
+        return undefined;
+      }
+    }
+
+    this.progress.startMonitoring();
+
+    this.sourceTileset = await loadFromArchive(
+      inputUrl,
+      I3SLoader,
+      {
+        ...this.loaderOptions,
+        // @ts-expect-error `isTileset` can be boolean of 'auto' but TS expects a string
+        i3s: {...this.loaderOptions.i3s, isTileset: true}
+      },
+      this.slpkFilesystem
+    );
 
     if (!this.sourceTileset) {
-      return;
+      return undefined;
     }
 
     const rootNode = this.sourceTileset?.root;
@@ -105,11 +137,28 @@ export default class Tiles3DConverter {
 
     this.tilesetPath = join(`${outputPath}`, `${tilesetName}`);
     this.attributeStorageInfo = this.sourceTileset.attributeStorageInfo;
+
+    await this.conversionDump.createDump(options as ConversionDumpOptions);
+    if (this.conversionDump.restored && this.options.inquirer) {
+      const result = await this.options.inquirer.prompt([
+        {
+          name: 'resumeConversion',
+          type: 'confirm',
+          message:
+            'Dump file of the previous conversion exists, do you want to resume that conversion?'
+        }
+      ]);
+      if (!result.resumeConversion) {
+        this.conversionDump.reset();
+      }
+    }
     // Removing the tilesetPath needed to exclude erroneous files after conversion
-    try {
-      await removeDir(this.tilesetPath);
-    } catch (e) {
-      // do nothing
+    if (!this.conversionDump.restored) {
+      try {
+        await removeDir(this.tilesetPath);
+      } catch (e) {
+        // do nothing
+      }
     }
 
     const rootTile: Tiles3DTileJSON = {
@@ -125,12 +174,54 @@ export default class Tiles3DConverter {
 
     const tileset = transform({root: rootTile}, tilesetTemplate());
     await writeFile(this.tilesetPath, JSON.stringify(tileset), 'tileset.json');
+    await this.conversionDump.deleteDumpFile();
 
-    this._finishConversion({slpk: false, outputPath, tilesetName});
+    this.progress.stopMonitoring();
+
+    await this._finishConversion({slpk: false, outputPath, tilesetName});
+
+    if (this.slpkFilesystem) {
+      this.slpkFilesystem.destroy();
+    }
 
     // Clean up worker pools
     const workerFarm = WorkerFarm.getWorkerFarm({});
     workerFarm.destroy();
+    return undefined;
+  }
+
+  /**
+   * Preprocess stage of the tile converter. Calculate number of nodes
+   * @returns true - the conversion is possible, false - the tileset's content is not supported
+   */
+  private async preprocessConversion(): Promise<boolean> {
+    // eslint-disable-next-line no-console
+    console.log('Analyze source layer');
+    const nodesCount = await getNodeCount(this.slpkFilesystem);
+    this.progress.stepsTotal = nodesCount;
+
+    // eslint-disable-next-line no-console
+    console.log('------------------------------------------------');
+    // eslint-disable-next-line no-console
+    console.log('Preprocess results:');
+    if (this.slpkFilesystem) {
+      // eslint-disable-next-line no-console
+      console.log(`Node count: ${nodesCount}`);
+      if (nodesCount === 0) {
+        // eslint-disable-next-line no-console
+        console.log('Node count is 0. The conversion will be interrupted.');
+        // eslint-disable-next-line no-console
+        console.log('------------------------------------------------');
+        return false;
+      }
+    } else {
+      // eslint-disable-next-line no-console
+      console.log('Node count cannot be calculated for the remote dataset');
+    }
+
+    // eslint-disable-next-line no-console
+    console.log('------------------------------------------------');
+    return true;
   }
 
   /**
@@ -140,15 +231,32 @@ export default class Tiles3DConverter {
    * @param level a current level of a tree depth
    * @param childNodeInfo child node to convert
    */
+  // eslint-disable-next-line complexity, max-statements
   private async convertChildNode(
     parentSourceNode: I3STileHeader,
     parentNode: Tiles3DTileJSON,
     level: number,
     childNodeInfo: NodeReference
   ): Promise<void> {
+    let nextParentNode = parentNode;
     const sourceChild = await this._loadChildNode(parentSourceNode, childNodeInfo);
     if (sourceChild.contentUrl) {
-      const content = await loadI3SContent(this.sourceTileset, sourceChild, this.loaderOptions);
+      if (
+        this.conversionDump.restored &&
+        this.conversionDump.isFileConversionComplete(`${sourceChild.id}.b3dm`) &&
+        (sourceChild.obb || sourceChild.mbs)
+      ) {
+        const {child} = this._createChildAndBoundingVolume(sourceChild);
+        parentNode.children.push(child);
+        await this._addChildren(sourceChild, child, level + 1);
+        return;
+      }
+      const content = await loadI3SContent(
+        this.sourceTileset,
+        sourceChild,
+        this.loaderOptions,
+        this.slpkFilesystem
+      );
 
       if (!content) {
         await this._addChildren(sourceChild, parentNode, level + 1);
@@ -162,39 +270,39 @@ export default class Tiles3DConverter {
         featureAttributes = await this._loadChildAttributes(sourceChild, this.attributeStorageInfo);
       }
 
-      if (!sourceChild.obb) {
-        sourceChild.obb = createObbFromMbs(sourceChild.mbs);
-      }
-
-      const boundingVolume = {
-        box: i3sObbTo3dTilesObb(sourceChild.obb, this.geoidHeightModel)
-      };
-      const child: Tiles3DTileJSON = {
-        boundingVolume,
-        geometricError: convertScreenThresholdToGeometricError(sourceChild),
-        children: []
-      };
+      const {child, boundingVolume} = this._createChildAndBoundingVolume(sourceChild);
 
       const i3sAttributesData: I3SAttributesData = {
         tileContent: content,
-        box: boundingVolume.box,
+        box: boundingVolume.box || [],
         textureFormat: sourceChild.textureFormat
       };
 
       const b3dmConverter = new B3dmConverter();
       const b3dm = await b3dmConverter.convert(i3sAttributesData, featureAttributes);
 
-      child.content = {
-        uri: `${sourceChild.id}.b3dm`,
-        boundingVolume
-      };
+      await this.conversionDump.addNode(`${sourceChild.id}.b3dm`, sourceChild.id);
       await writeFile(this.tilesetPath, new Uint8Array(b3dm), `${sourceChild.id}.b3dm`);
+      await this.conversionDump.updateConvertedNodesDumpFile(
+        `${sourceChild.id}.b3dm`,
+        sourceChild.id,
+        true
+      );
       parentNode.children.push(child);
-
-      await this._addChildren(sourceChild, child, level + 1);
-    } else {
-      await this._addChildren(sourceChild, parentNode, level + 1);
+      nextParentNode = child;
     }
+
+    this.progress.stepsDone += 1;
+    let timeRemainingString = 'Calculating time left...';
+    const timeRemaining = this.progress.getTimeRemainingString();
+    if (timeRemaining) {
+      timeRemainingString = `${timeRemaining} left`;
+    }
+    const percentString = this.progress.getPercentString();
+    const progressString = percentString ? ` ${percentString}%, ${timeRemainingString}` : '';
+    console.log(`[converted${progressString}]: ${childNodeInfo.id}`); // eslint-disable-line
+
+    await this._addChildren(sourceChild, nextParentNode, level + 1);
   }
 
   /**
@@ -235,18 +343,46 @@ export default class Tiles3DConverter {
     } else {
       const nodeUrl = this._relativeUrlToFullUrl(parentNode.url, childNodeInfo.href!);
       // load metadata
-      const options = {
+      const options: I3SLoaderOptions = {
         i3s: {
           ...this.loaderOptions,
+          // @ts-expect-error
           isTileHeader: true,
           loadContent: false
         }
       };
 
       console.log(`Node conversion: ${nodeUrl}`); // eslint-disable-line no-console,no-undef
-      header = await load(nodeUrl, I3SLoader, options);
+      header = await loadFromArchive(nodeUrl, I3SLoader, options, this.slpkFilesystem);
     }
     return header;
+  }
+
+  /**
+   * Create child and child's boundingVolume for the converted node
+   * @param sourceChild
+   * @returns child and child's boundingVolume
+   */
+  private _createChildAndBoundingVolume(sourceChild: I3STileHeader): {
+    boundingVolume: Tile3DBoundingVolume;
+    child: Tiles3DTileJSON;
+  } {
+    if (!sourceChild.obb) {
+      sourceChild.obb = createObbFromMbs(sourceChild.mbs);
+    }
+    const boundingVolume: Tile3DBoundingVolume = {
+      box: i3sObbTo3dTilesObb(sourceChild.obb, this.geoidHeightModel)
+    };
+    const child: Tiles3DTileJSON = {
+      boundingVolume,
+      geometricError: convertScreenThresholdToGeometricError(sourceChild),
+      children: [],
+      content: {
+        uri: `${sourceChild.id}.b3dm`,
+        boundingVolume
+      }
+    };
+    return {boundingVolume, child};
   }
 
   /**
@@ -292,7 +428,7 @@ export default class Tiles3DConverter {
         attributeType: this._getAttributeType(attribute)
       };
 
-      promises.push(load(inputUrl, I3SAttributeLoader, options));
+      promises.push(loadFromArchive(inputUrl, I3SAttributeLoader, options, this.slpkFilesystem));
     }
     const attributesList = await Promise.all(promises);
     this._replaceNestedArrays(attributesList);
